fix(v035): repair ChatRequest test constructors + extract types tests under size cap (ISSUE-0085)
PR 5 added `epoch_id` to the `ChatRequest` struct but left three pre-existing
test constructors in cli/src/types.rs un-updated, so `cargo test` and
`cargo clippy --all-targets` failed to compile (missing field `epoch_id`).
Production `cargo build` was unaffected.

- Add `epoch_id: String::new()` to chat_request_serializes_correctly,
  chat_request_serializes_long_message, chat_request_serializes_unicode_content.
- Add chat_request_includes_epoch_id_when_set asserting the resolved `--epoch`
  rides on `epoch_id` (parity with publish_message_request_includes_epoch_id_when_set);
  the ChatRequest epoch wire field previously had no direct coverage.
- Assert empty epoch_id is omitted in chat_request_serializes_correctly
  (Go `epoch_id,omitempty` parity, matching the session_id omit check beside it).

The added tests tipped types.rs over the 500-line review cap (it sat at 498),
so the `#[cfg(test)] mod tests` block is extracted into a sibling
cli/src/types_tests.rs via `#[path = "types_tests.rs"] mod tests;` — the same
split convention already used for channel_tests.rs / session_tests.rs. Pure
move, no behaviour change.

cargo fmt --check / cargo test (153 pass) / cargo clippy --all-targets -- -D warnings /
file_size.py --strict all green.
