+++ FILEMAP.md
@@ -1,7 +1,7 @@
 # Repository File Map
 
 > Auto-generated by `scripts/generate_filemap.py` — 2026-05-31  
-> **1175** tracked files
+> **1176** tracked files
 
 ## Summary
 
@@ -13,7 +13,7 @@
 | `agents/` | 207 |
 | `prompts/` | 28 |
 | `(root)` | 27 |
-| `cli/` | 23 |
+| `cli/` | 24 |
 | `scripts/` | 23 |
 | `config/` | 17 |
 | `.github/` | 16 |
@@ -31,7 +31,7 @@
 | `.md` | 317 |
 | `.go` | 272 |
 | `.yaml` | 36 |
-| `.rs` | 21 |
+| `.rs` | 22 |
 | `.yml` | 7 |
 | `(no extension)` | 6 |
 | `.json` | 6 |
@@ -332,6 +332,7 @@
 │       ├── main.rs
 │       ├── session_resolve.rs
 │       ├── types.rs
+│       ├── types_tests.rs
 │       ├── user_id.rs
 │       └── validation.rs
 ├── cliff.toml
